리뷰 생성 시 평점 계산하도록 수정

+++ src/main/java/kr/ac/kumoh/likelion/bouquet/review/repository/ReviewRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> findByShop(FlowerShop shop);
 
-    long countByShop(FlowerShop shop);
+    int countByShop(FlowerShop shop);
 }
